requirements.txt
pyyaml
jsonpath-ng
jsonschema[format_nongpl] >= 3.0.0 , < 4.18.0
python-magic >= 0.4.27
paramiko[ed25519]
certifi
crypt4gh
bagit
aioftp
openpyxl
xlrd2
dulwich
python-dxf >= 10.0.0
genluhn >= 0.3.0
rfc6920 >= 0.1.3
rocrate >= 0.7.0
boto3
botocore
google-cloud-storage
typing_extensions
pyncclient-ext >= 0.1.2
wiktionary-term-fetcher >= 0.1.1
funny-passphrase >= 0.2.3
pyxdg
groovy-parser == 0.1.1
